requirements.txt
torch
pyyaml
matplotlib
numpy
streamlit
